Improve error handling
This both makes the error handling a bit nicer internally and makes it
so the cache is always exported (unless we segfault or something)
