+++ cmd/print_vars.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"os"
 
 	"github.com/spf13/cobra"
 
@@ -10,7 +9,7 @@ import (
 	"github.com/redbubble/yak/cli"
 )
 
-func printVarsCmd(cmd *cobra.Command, args []string) {
+func printVarsCmd(cmd *cobra.Command, args []string) error {
 	roleName := cli.ResolveRole(args[0])
 
 	creds := cli.AssumeRoleFromCache(roleName)
@@ -20,16 +19,14 @@ func printVarsCmd(cmd *cobra.Command, args []string) {
 		loginData, err := cli.GetLoginData()
 
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "%v\n", err)
-			os.Exit(1)
+			return err
 		}
 
 		cli.CacheLoginRoles(loginData.Roles)
 		creds, err = cli.AssumeRole(loginData, roleName)
 
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "%v\n", err)
-			os.Exit(1)
+			return err
 		}
 
 		cli.CacheCredentials(roleName, creds)
@@ -38,4 +35,6 @@ func printVarsCmd(cmd *cobra.Command, args []string) {
 	for key, value := range aws.EnvironmentVariables(creds.Credentials) {
 		fmt.Printf("export %s='%s'\n", key, value)
 	}
+
+	return nil
 }
